test: Moved unit tests for MVEM with gravity.
